Extended Sunshine so SPT could be ported.
The changes in the sunshine.util package should probably be moved
to somewhere else.
But for now they suffice.
Also, I'm unsure if there are ever several ReferenceRule entries in
packed.esv.
Finally, I'm now depending on the fact that the analysisFunction
of a language is actually the observer strategy.